# Simplify and speed up three Doppler equivalencies