### Description

I've managed to speed up creating the `doppler_radio()`, `doppler_optical()` and `doppler_relativistic()` equivalencies by almost an order of magnitude.

Setup:
```
Python 3.12.3 (main, Feb  4 2025, 14:48:35) [GCC 13.3.0]
Type 'copyright', 'credits' or 'license' for more information
IPython 9.1.0 -- An enhanced Interactive Python. Type '?' for help.
Tip: You can use Ctrl-O to force a new line in terminal IPython

In [1]: from astropy import units as u

In [2]: CO_restfreq = 115.27120*u.GHz
```

On current `main`:
```
In [3]: %timeit u.doppler_radio(CO_restfreq)
58.7 μs ± 88.1 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

In [4]: %timeit u.doppler_optical(CO_restfreq)
59 μs ± 257 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)

In [5]: %timeit u.doppler_relativistic(CO_restfreq)
59.4 μs ± 286 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

With this patch:
```
In [3]: %timeit u.doppler_radio(CO_restfreq)
6.64 μs ± 21.1 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

In [4]: %timeit u.doppler_optical(CO_restfreq)
6.71 μs ± 25 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

In [5]: %timeit u.doppler_relativistic(CO_restfreq)
5.78 μs ± 7.83 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)
```

There is also a `doppler_redshift()` equivalency, but it is quite different from the other three because it has no `rest` parameter, which in turn has allowed it to be cached.

- [x] By checking this box, the PR author has requested that maintainers do **NOT** use the "Squash and Merge" button. Maintainers should respect this when possible; however, the final decision is at the discretion of the maintainer that merges the PR.